fix: restructure routing
