+++ src/controllers/webhook/ai_hub/index.js
@@ -1,5 +1,5 @@
-import verifyAIHubWebhook from "../../../../auth/aihub-auth";
-import LeadController from "../lead";
+import verifyAIHubWebhook from "../../../auth/aihub-auth";
+import LeadController from "../erp/lead"
 
 export default class AIHubWebhook {
   static register(webhook) {
